Refactor comparing task: move assert_valid_run function
